+++ pytorch_lightning/loops/base.py
@@ -18,6 +18,8 @@ from typing import Any, Dict, Optional
 from deprecate import void
 
 import pytorch_lightning as pl
+from pytorch_lightning.trainer.progress import BaseProgress, Tracker
+from pytorch_lightning.utilities.apply_func import apply_to_collection
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
 
@@ -45,8 +47,35 @@ class Loop(ABC):
 
     def __init__(self) -> None:
         self.iteration_count: int = 0
-        self.trainer: Optional['pl.Trainer'] = None
+        self._trainer: Optional['pl.Trainer'] = None
         self.restarting = False
+
+    @property
+    def loop_progress(self) -> Dict[str, Any]:
+        """Return the progress for the current loop and children loop."""
+        progress = {}
+        for k, v in self.__dict__.items():
+            if isinstance(v, BaseProgress):
+                progress[k] = v
+            elif isinstance(v, Loop):
+                progress[k] = v.loop_progress
+        return progress
+
+    @property
+    def trainer(self) -> Optional['pl.Trainer']:
+        return self._trainer
+
+    @trainer.setter
+    def trainer(self, trainer: 'pl.Trainer'):
+        """Connect the Trainer to itself and all its children loops"""
+        if not isinstance(trainer, pl.Trainer):
+            raise MisconfigurationException(
+                f"Loop {self.__class__.__name__} should be connected to a `Trainer`, found: {trainer}."
+            )
+        self._trainer = trainer
+        for v in self.__dict__.values():
+            if isinstance(v, Loop):
+                v.trainer = trainer
 
     @property
     @abstractmethod
@@ -61,10 +90,6 @@ class Loop(ABC):
     def connect(self, trainer: 'pl.Trainer', *args: Any, **kwargs: Any) -> None:
         """Connects Loop with all the necessary things like connectors and accelerators."""
         # TODO(@justusschock): Make the trainer a weakref/proxy
-        if not isinstance(trainer, pl.Trainer):
-            raise MisconfigurationException(
-                f"Loop {self.__class__.__name__} should be connected to a `Trainer`, found: {trainer}."
-            )
         self.trainer = trainer
 
     def on_skip(self) -> Optional[Any]:
@@ -85,14 +110,13 @@ class Loop(ABC):
         Returns:
             the output of :attr:`on_run_end` (often outputs collected from each step of the loop)
         """
+        if self.trainer is None:
+            raise MisconfigurationException(f"The {self.__class__.__name__} Loop hasn't been attached to any Trainer.")
+
         if self.skip:
             return self.on_skip()
 
-        if self.restarting:
-            self.restore()
-            self.restarting = False
-        else:
-            self.reset()
+        self.reset()
 
         self.on_run_start(*args, **kwargs)
 
@@ -107,9 +131,6 @@ class Loop(ABC):
 
         output = self.on_run_end()
         return output
-
-    def restore(self) -> None:
-        """Restore the internal state of the loop the beginning of run if restarting is ``True``."""
 
     @abstractmethod
     def reset(self) -> None:
@@ -142,9 +163,50 @@ class Loop(ABC):
     def teardown(self) -> None:
         """Use to release memory etc."""
 
-    def load_state_dict(self, state_dict: Dict) -> None:
-        """Restore the loop state from the provided state_dict."""
+    def on_save_checkpoint(self) -> Dict:
+        """
+        Called when saving a model checkpoint, use to persist loop state.
 
-    def state_dict(self) -> Dict:
-        """Return the loop current states."""
+        Returns:
+            The current loop state.
+        """
         return {}
+
+    def on_load_checkpoint(self, state_dict: Dict):
+        """Called when loading a model checkpoint, use to reload loop state."""
+
+    def state_dict(self, destination: Optional[Dict] = None, prefix: Optional[str] = '') -> Dict:
+        if destination is None:
+            destination = {}
+
+        destination[prefix + "state_dict"] = self.on_save_checkpoint()
+
+        for k, v in self.__dict__.items():
+            if isinstance(v, BaseProgress):
+                destination[prefix + k] = v.state_dict()
+            elif isinstance(v, Loop):
+                v.state_dict(destination, prefix + k + '.')
+        return destination
+
+    def _load_from_state_dict(self, state_dict, prefix, restart_progress):
+        for k, v in self.__dict__.items():
+            if isinstance(v, BaseProgress):
+                v.load_state_dict(state_dict[prefix + k])
+                if restart_progress:
+
+                    def restart(v: Tracker):
+                        v.reset_on_restart()
+
+                    apply_to_collection(v, Tracker, restart)
+
+        self.on_load_checkpoint(state_dict[prefix + "state_dict"])
+        self.restarting = True
+
+    def __load(self, state_dict, restart_progress, prefix=''):
+        self._load_from_state_dict(state_dict, prefix, restart_progress)
+        for k, v in self.__dict__.items():
+            if isinstance(v, Loop):
+                v.__load(state_dict.copy(), restart_progress, prefix + k + '.')
+
+    def load_state_dict(self, state_dict: Dict, restart_progress: bool = True):
+        self.__load(state_dict.copy(), restart_progress)